requirements.txt
customtkinter
Pillow
python-docx
pymupdf
pandas
openpyxl